Moved the plotting function to ./helpers/ and refactored the code
I used this plotting script a few times in different parts of the project and probably will again, so it should be in its own folder. This helpers folder can also be used to contain other helpers I may need.

The plotting_functions.jl was refactored to account for some overlapping code and now is more modular.