Added docstrings and experimental training set plotting to the phase plane.
I didn't test this new code as I was generating frames for a different test at the same time. The phase plane plot should be changed to graph the training set in yellow, but i did not test this. zorder is changed so that it plots on top of the blue one, but this may not work.

Maybe I should consider changing the static_struct so that the testing time is included with the training time as Vector{Float64}s. The true solution should always be plotted on the test set, then additional functionality should be included to indicate where the training region is. Thus we should change observation data to be a name which reflects this. Observations are training sets, not test sets.