activation getter is complete and the code is refactored to be self contained. Hiddenodes-1 currently plots the activation fucntions.
There is now support for getting the activation functions of the KAN and ive also included a simple function to plot it.

I took some scripts that were previously in the KANODEs folder and put it in the hidden-odes folder so everything is self contained. I removed the GPU code as it is so far behind everything else that it is better to just to it seperately later on.